Add comments to variables used inside of action constructors. Delete Direct_Tick_Callback as it no longer works, and special_proc as it wasn't used anywhere. Fix a few compiler warnings.

+++ engine/class_modules/sc_hunter.cpp
@@ -2081,7 +2081,7 @@ struct black_arrow_t: public hunter_ranged_attack_t
     while ( proc_count > 0 )
     {
       // note that ticks start from 1
-      int proc_tick = floor( p() -> rng().range( 0, possible_slots ) ) + 1;
+      int proc_tick = static_cast<int>( floor( p() -> rng().range( 0, possible_slots ) ) + 1 );
 
       // Given N is the proc_tick, find the Nth tick that is not yet proccing
       uint32_t proc_bit = 1;
